+++ backend/src/models/__init__.py
@@ -107,6 +107,7 @@ class Chat(BaseModel):
     streaming: bool = False
     memory: str = MemoryType.CONVERSATION_KG
     model: str = ModelType.OPENAI_GPT_3_5_TURBO_16K
+    system: str = None
     messages: list
     
     __config__ = {
@@ -115,8 +116,9 @@ class Chat(BaseModel):
 				"streaming": False,
 				"memory": MemoryType.CONVERSATION_KG,
                 "model": ModelType.OPENAI_GPT_3_5_TURBO_16K,
+                "system": None,
 				"messages": [
-					{"role": "system", "content": "You are a helpful assistant."},
+                    {"role": "system", "content": "You are a helpful assistant."},
 					{"role": "user", "content": 'Who won the 2001 world series?'},
 					{"role": "assistant", "content": 'The arizona diamondbacks won the 2001 world series.'},
 					{
@@ -169,8 +171,8 @@ class Retrieval(Chat):
 	}
     
 class Agent(Chat):
-    tools: list[str]
-    retrieval: RetrievalParams
+    tools: Optional[List[str]] = None
+    retrieval: Optional[RetrievalParams] = None
     
     __config__ = {
 		"json_schema_extra": {
@@ -192,7 +194,7 @@ class Agent(Chat):
 				],
                 "tools": ["repl_tool", "csv_tool"],
                 "retrieval": {
-                    "provider": SearchProvider.PINECONE,
+                    "provider": SearchProvider.REDIS,
                     "index_name": "",
                     "search_type": SearchType.SIMILARITY,
                     "search_kwargs": {