+++ backend/src/repositories/chat.py
@@ -1,10 +1,13 @@
+import ujson
+import hashlib
 from datetime import datetime
 
 from sqlalchemy import select
 from sqlalchemy.orm import joinedload
 
-from src.models import Chat as ChatBody
+from src.models import Agent as ChatBody
 from src.models.sql.chat import Chat, Image, Message, Source
+from src.utils.format import hash_string
 
 class ChatRepository:
     def __init__(self, request, db):
@@ -14,7 +17,9 @@ class ChatRepository:
     async def list(self):
         stmt = (
             select(Chat)
-            .options(joinedload(Chat.messages).load_only(Message.id, Message.role, Message.content, Message.model, Message.created_at))
+            .options(joinedload(Chat.messages).load_only(
+                Message.id, Message.role, Message.content, Message.model, Message.created_at
+            ))
             .where(Chat.user_id == self.user_id)
             .order_by(Chat.updated_at.desc())
         )
@@ -26,10 +31,13 @@ class ChatRepository:
         chats_json = [
             {
                 "id": chat.id,
+                "system": chat.system or None,
                 "messages": [
                     {"role": message.role, "content": message.content, "model": message.model, "created_at": message.created_at.isoformat()}
                     for message in sorted(chat.messages, key=lambda m: m.created_at)  # Sort messages here if not sorted in model
                 ],
+                "retrieval": ujson.loads(chat.retrieval) if chat.retrieval else None,
+                "tools": ujson.loads(chat.tools) if chat.tools else None,
                 "created_at": chat.created_at.isoformat(),
                 "updated_at": chat.updated_at.isoformat(),
             }
@@ -41,7 +49,10 @@ class ChatRepository:
     async def create(self, chat: ChatBody):
         async with self.db.begin() as transaction:
             try:
-                new_chat = Chat(user_id=self.user_id, 
+                new_chat = Chat(user_id=self.user_id,
+                                system=chat.system,
+                                retrieval=chat.retrieval.model_dump_json(),
+                                tools=ujson.dumps(chat.tools), 
                                 created_at=datetime.utcnow(), 
                                 updated_at=datetime.utcnow())
                 self.db.add(new_chat)
@@ -96,6 +107,7 @@ class ChatRepository:
         if chat:
             chat_json = {
                 "id": chat.id,
+                "system": chat.system or None,
                 "messages": [
                     {
                         "role": message.role, 
@@ -116,6 +128,8 @@ class ChatRepository:
                     }
                     for message in chat.messages
                 ],
+                "retrieval": ujson.loads(chat.retrieval) if chat.retrieval else None,
+                "tools": ujson.loads(chat.tools) if chat.tools else None,
                 "created_at": chat.created_at.isoformat(),
                 "updated_at": chat.updated_at.isoformat(),
             }
@@ -137,6 +151,27 @@ class ChatRepository:
 
                     # Set the updated_at field
                     chat.updated_at = datetime.utcnow()
+                    
+                    # Update tools, system, and retrieval if present in updates
+                    if updates.system is not None:
+                        if chat.system != updates.system:
+                            chat.system = updates.system
+                    
+                    if updates.tools is not None:
+                        new_tools_json = ujson.dumps(updates.tools)
+                        current_tools_hash = hash_string(chat.tools)
+                        new_tools_hash = hash_string(new_tools_json)
+                        
+                        if current_tools_hash != new_tools_hash:
+                            chat.tools = new_tools_json
+                        
+                    if updates.retrieval is not None:
+                        new_retrieval_json = updates.retrieval.model_dump_json()
+                        current_retrieval_hash = hash_string(chat.retrieval)
+                        new_retrieval_hash = hash_string(new_retrieval_json)
+
+                        if current_retrieval_hash != new_retrieval_hash:
+                            chat.retrieval = new_retrieval_json
 
                     # Add new messages
                     messages = updates.messages