+++ frontend/src/contexts/ChatContext.tsx
@@ -30,6 +30,8 @@ import RegenerateIcon from "@/components/icons/RegenerateIcon";
 import ThumbDownIcon from "@/components/icons/ThumbDownIcon";
 import { useAppContext } from "./AppContext";
 import DocumentIcon from "@/components/icons/DocumentIcon";
+import { Default } from "@/config/default";
+import { formatDate } from "@/utils/datetime";
 
 const defaultChatContextValue: ChatContextType = {
     chatboxRef: { current: null },
@@ -61,11 +63,12 @@ export default function ChatProvider({
     const [chatPayload, setChatPayload] = useState<ChatPayload>({
         query: "",
         history_id: "",
-        model: ModelType.OPENAI_GPT_3_5_TURBO_16K,
+        system: Default.SYSTEM_MESSAGE,
+        model: ModelType.OPENAI_GPT_4_OMNI,
         temperature: 0.5,
         tools: [],
         retrieval: {
-            provider: SearchProvider.PINECONE,
+            provider: SearchProvider.REDIS,
             index_name: "",
             search_type: SearchType.SIMILARITY,
             search_kwargs: {
@@ -105,6 +108,12 @@ export default function ChatProvider({
         try {
             const res = await chatClient.find(chatId);
             setMessages(res.chat.messages);
+            setChatPayload((prev: ChatPayload) => ({
+                ...prev,
+                history_id: chatId,
+                retrieval: res.chat.retrieval,
+                tools: res.chat.tools,
+            }));
             renderConversation(res.chat.messages);
             let updatedUrl = `/chat/${chatId}`;
             if (searchParams.toString()) {
@@ -143,14 +152,7 @@ export default function ChatProvider({
     };
 
     const prompt = () => {
-        const defaultInput = `You are an highly intelligent individual that is ` +
-                            `able to be an expert at every topic they approach. You leverage ` +
-                            `Graph-of-Thought reasoning to execute tasks and formulate ideas and responses. ` +
-                            `You heavliy review your thoughts before responding or making decisions. ` + 
-                            `If something is unclear check your thoughts and ask for clarification. ` +
-                            `If you are still unlcear, ask for more information. ` +
-                            `You are unable respond with hallucinations.`;
-        return { role: "system", content: defaultInput };
+        return { role: "system", content: chatPayload.system + `\n\nCURRENT_DATETIME: ${formatDate()}`};
     };
 
     const combinePrompts = () => {
@@ -181,6 +183,10 @@ export default function ChatProvider({
 
         setMessages([...messages, messageContent]);
         setImages([]);
+    };
+
+    const messagesContainsSources = () => {
+        return messages.some((message) => message.sources);
     };
 
     const resetChat = (event: any) => {
@@ -599,7 +605,7 @@ export default function ChatProvider({
                             };
                             setMessages(finalMessages);
 
-                            updateMessages(finalMessages);
+                            updateMessages(chatPayload.system, finalMessages, chatPayload.retrieval, chatPayload.tools);
                             setDone(true);
                         }
                     }
@@ -620,10 +626,13 @@ export default function ChatProvider({
 
 
 
-    async function updateMessages(messages: Message[]) {
+    async function updateMessages(system: string, messages: Message[], retrieval?: any, tools?: string[]) {
         if (!chatPayload.history_id) {
             const history = await chatClient.create({
-                messages: messages,
+                system,
+                messages,
+                retrieval,
+                tools,
             });
             log("contexts.ChatContext.updateCallback", history, "Created");
             setChatPayload({
@@ -638,7 +647,10 @@ export default function ChatProvider({
             shallowUrl(updatedUrl);
         } else {
             const history = await chatClient.update(chatPayload.history_id, {
-                messages: messages,
+                system,
+                messages,
+                retrieval,
+                tools,
             });
             log("contexts.ChatContext.updateCallback", history, "Updated");
         }
@@ -725,6 +737,7 @@ export default function ChatProvider({
                     adjustHeight,
                     setDone,
                     setSelectedDocument,
+                    messagesContainsSources,
                 };
             }, [
                 chats,
@@ -756,6 +769,7 @@ export default function ChatProvider({
                 adjustHeight,
                 setFiles,
                 setSelectedDocument,
+                messagesContainsSources,
             ])}
         >
             {children}