+++ Changelog.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## v0.0.1
 ### Fixed
+  - bugfix/42-redis-is-default-vector-db (2024-05-26)
   - bugfix/33-groq-rag-not-working (2024-05-19)
 ### Changed 
+  - feature/46-add-system-to-chat-record (2024-05-26)
+  - feature/44-can-create-index-from-existing (2024-05-26)
   - feature/35-db-indexes (2024-05-19)
   - feature/23-message-include-model-used (2024-05-19)
   - feature/29-storing-sources-on-messages (2024-05-19)